+++ opentelemetry-semantic-conventions/src/opentelemetry/semconv/schemas.py
@@ -90,5 +90,15 @@ class Schemas(Enum):
     The URL of the OpenTelemetry schema version 1.38.0.
     """
 
+    V1_39_0 = "https://opentelemetry.io/schemas/1.39.0"
+    """
+    The URL of the OpenTelemetry schema version 1.39.0.
+    """
+
+    V1_40_0 = "https://opentelemetry.io/schemas/1.40.0"
+    """
+    The URL of the OpenTelemetry schema version 1.40.0.
+    """
+
     # when generating new semantic conventions,
     # make sure to add new versions version here.